+++ toolkit/offchain/src/plutus_script.rs
@@ -70,3 +70,42 @@ impl PlutusScript {
 		PolicyId(self.script_hash())
 	}
 }
+
+#[cfg(test)]
+pub(crate) mod tests {
+	use super::*;
+	use chain_params::SidechainParams;
+	use hex_literal::hex;
+	use sidechain_domain::{MainchainAddressHash, McTxHash, UtxoId, UtxoIndex};
+
+	pub(crate) const TEST_PARAMS: SidechainParams = SidechainParams {
+		chain_id: 111,
+		threshold_numerator: 2,
+		threshold_denominator: 3,
+		genesis_committee_utxo: UtxoId {
+			tx_hash: McTxHash(hex!(
+				"0000000000000000000000000000000000000000000000000000000000000000"
+			)),
+			index: UtxoIndex(0),
+		},
+		governance_authority: MainchainAddressHash(hex!(
+			"76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"
+		)),
+	};
+
+	// Taken from smart-contracts repository
+	pub(crate) const CANDIDATES_SCRIPT_RAW: [u8; 318] = hex!("59013b590138010000323322323322323232322222533553353232323233012225335001100f2215333573466e3c014dd7001080909802000980798051bac330033530040022200148040dd7198011a980180311000a4010660026a600400644002900019112999ab9a33710002900009805a4810350543600133003001002300f22253350011300b49103505437002215333573466e1d20000041002133005337020089001000919199109198008018011aab9d001300735573c0026ea80044028402440204c01d2401035054350030092233335573e0024016466a0146ae84008c00cd5d100124c6010446666aae7c00480288cd4024d5d080118019aba20024988c98cd5ce00080109000891001091000980191299a800880211099a80280118020008910010910911980080200191918008009119801980100100081");
+
+	/// We know it is correct, because we are able to get the same hash as using code from smart-contract repository
+	pub(crate) const CANDIDATES_SCRIPT_WITH_APPLIED_PARAMS: [u8; 400] = hex!("59018d0100003323322323322323232322222533553353232323233012225335001100f2215333573466e3c014dd7001080909802000980798051bac330033530040022200148040dd7198011a980180311000a4010660026a600400644002900019112999ab9a33710002900009805a490350543600133003001002300f22253350011300b49103505437002215333573466e1d20000041002133005337020089001000919199109198008018011aab9d001300735573c0026ea80044028402440204c01d2401035054350030092233335573e0024016466a0146ae84008c00cd5d100124c6010446666aae7c00480288cd4024d5d080118019aba20024988c98cd5ce00080109000891001091000980191299a800880211099a802801180200089100109109119800802001919180080091198019801001000a60151d8799f186fd8799fd8799f58200000000000000000000000000000000000000000000000000000000000000000ff00ff0203581c76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9ff0001");
+
+	#[test]
+	fn apply_parameters_to_deregister() {
+		let applied =
+			PlutusScript::from_wrapped_cbor(&CANDIDATES_SCRIPT_RAW, LanguageKind::PlutusV2)
+				.unwrap()
+				.apply_data(TEST_PARAMS)
+				.unwrap();
+		assert_eq!(hex::encode(applied.bytes), hex::encode(CANDIDATES_SCRIPT_WITH_APPLIED_PARAMS));
+	}
+}
